Add custom request handler for making external requests.
When WP uses the nonce middleware, that header is added to every request no matter
what. Requests to external install will always fail if that header is present.

To support this, I've introduced a custom request handler with an `addMiddleware`
function which runs middleware right before the request is made (after WP core
middlewares are run). This allows things like removing the nonce headers and/or
setting the root URL.