Fix delay when rendering styles on front-end
We were waiting 2 seconds for the dom to load before rendering styles
which cause fast sites to have a delay before being styled on the front
end.

We not add the styles to the <head> as well as inside the iframe when
using Gutenberg.

No more delay on non Gutenberg and still support gutenberg.