If gutenerg iframe not found in 2 seconds try again 5 seconds later
If PHP Debug is on there is typically a delay in loading the iframe. If
the iframe is not found in time styles would not work as the iframe was
marked as 'no-iframe-available'.

Now if we don't find the iframe the first time, we try again 5 seconds
later and mark 'no-iframe-available' if not found.

Using 2 timers instead of one 7-second timer to avoid extra delays if
the iframe loaded quickly.